Copy Memory to Acceleration Structure
This commit adds the last copy command, that being copy memory to acceleration structure. This command is almost the same as the other copy commands, and I'm getting close to finishing up this PR.

+++ MoltenVK/MoltenVK/GPUObjects/MVKDevice.mm
@@ -981,7 +981,7 @@ void MVKPhysicalDevice::getProperties(VkPhysicalDeviceProperties2* properties) {
                 accelerationStructureProps->maxGeometryCount = pow(2, 24);
                 accelerationStructureProps->maxInstanceCount = pow(2, 24);
                 accelerationStructureProps->maxPrimitiveCount = pow(2, 28);
-                accelerationStructureProps->minAccelerationStructureScratchOffsetAlignment = _metalFeatures.mtlBufferAlignment;
+                accelerationStructureProps->minAccelerationStructureScratchOffsetAlignment = (uint32_t)_metalFeatures.mtlBufferAlignment;
                 // Other properties have not been figured out quite yet
                 break;
             }
